[ADD] Improved support for Odoo > 15.0
We added support for the specific parameters added in Odoo 16.0, as the
longpoll port have been deprecated.

The container type was changed from longpoll to chat as this is the
given name by Odoo even in the previous versions.

Renamed the variable for the container type to
`ORCHESTSH_CONTAINER_TYPE` to match the rest of the variables used by
OrchestSh to customize the container behavior.

Updated Go version so we can use the latest version of semver to check
which version we are using.