[FIX] Autostart configuration
The autostart had a bug where it was setting the variable to false
always no matter what value it had (as log as it had a value).

Added a second Setter to force the autostart value of the Odoo instance
in particular. This solves the issue where the variable wasn't updated
when we execute more complicated workflows (mainly initialization) where
we restart the container with different values for AUTOSTART.